move count function to aggregate functions
